requirements.txt
numpy==1.26.2
pandas==2.1.4
tensorflow==2.15.0
keras==2.15.0
pillow==10.2.0
scikit-learn==1.3.2